feat: add vote handling service, improve validation, and optimize queries

+++ geeksclub-api/src/main/java/dev/sivalabs/geeksclub/rest/dto/MessageDetailResponse.java
@@ -18,5 +18,5 @@ public record MessageDetailResponse(
 
     public record AuthorInfo(Long id, String fullName, String username) {}
 
-    public record VotesInfo(int upvoteCount, int downvoteCount, int score) {}
+    public record VotesInfo(long upvoteCount, long downvoteCount, long score) {}
 }
