+++ geeksclub-api/src/main/java/dev/sivalabs/geeksclub/rest/dto/MessageFeedItem.java
@@ -15,5 +15,5 @@ public record MessageFeedItem(
 
     public record AuthorInfo(Long id, String fullName, String username) {}
 
-    public record VotesInfo(int upvoteCount, int downvoteCount, int score) {}
+    public record VotesInfo(long upvoteCount, long downvoteCount, long score) {}
 }
