+++ geeksclub-api/src/main/java/dev/sivalabs/geeksclub/rest/dto/VoteResponse.java
@@ -4,5 +4,5 @@ import dev.sivalabs.geeksclub.domain.dto.VoteType;
 import java.time.Instant;
 
 public record VoteResponse(Long messageId, VoteType voteType, Votes votes, Instant votedAt) {
-    public record Votes(int upvoteCount, int downvoteCount, int score) {}
+    public record Votes(long upvoteCount, long downvoteCount, long score) {}
 }
